pyproject.toml
[tool.poetry]
name = "twock"
version = "0.1.2"
description = "Ping/Knock a list of tweets to enquire their status (deleted/available/...) #twockknock"
authors = ["Felix Victor Münch <f.muench@leibniz-hbi.de>","Philipp Kessling <p.kessling@leibniz-hbi.de>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Leibniz-HBI/twock"
classifiers = [
        "Development Status :: 3 - Alpha",
        "Environment :: Console",
        "Intended Audience :: Science/Research",
]

[tool.poetry.dependencies]
python = "^3.8"
twarc = "^2.3.10"
pandas = "^1.3.1"
PyYAML = "^5.4.1"
ndjson = "^0.3.1"
click = "^8.0.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyterlab = "^3.1.1"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
twock = 'twock.twock:cli'
